Struct my_rust_lib::boot_node::BootNode

pub struct BootNode<'a> {
    pub(crate) session: Arc<Session>,
    pub(crate) zid: String,
    pub(crate) cluster_name: String,
    pub(crate) received_counter: i32,
    pub(crate) expected_counter: i32,
    sub_boot: Subscriber<'a, Receiver<Sample>>,
    sub_counter: Subscriber<'a, Receiver<Sample>>,
    pub cluster: OrderedMapPairs,
    pub polygon_list: OrderedMapPolygon,
    pub correct_polygon_list: OrderedMapPolygon,
    pub draw_count: i32,
}

A boot node in a network acts as the entry point for new nodes to join the cluster. It also acts as a central point for nodes to leave the cluster. Stores the site and polygon information of all nodes in the network. Constructs the distributed voronoi diagram from received polygons as well as a centralized voronoi diagram.

Fields
session: Arc<Session>zid: Stringcluster_name: Stringreceived_counter: i32expected_counter: i32sub_boot: Subscriber<'a, Receiver<Sample>>sub_counter: Subscriber<'a, Receiver<Sample>>cluster: OrderedMapPairspolygon_list: OrderedMapPolygoncorrect_polygon_list: OrderedMapPolygondraw_count: i32
Implementations

impl BootNode<'_>

pub fn new(cluster_name: &str) -> Self

Creates a new boot node instance with a session. Opens a subscription on topic {cluster}/boot/* to receive incoming messages from nodes and a subscription on{cluster}/counter/* to count the number of messages its received since processing the current message on the {cluster}/boot/* topic.

pub fn run(&mut self)

Process the current messages that are in the subscription channel queue one at a time. Handles each topic with a different handler.

pub fn get_zid(&self) -> String

Get the zid of the node