Module my_rust_lib::node

Modules

config

Configuration to pass to zenoh::open() and zenoh::scout() functions and associated constants.

Structs

Config

The main configuration structure for Zenoh.

KeyExpr

A possibly-owned version of keyexpr that may carry optimisations for use with a Session that may have declared it.

Locator

A String that respects the Locator canon form: <proto>/<address>?<metadata>, such that <metadata> is of the form <key1>=<value1>;...;<keyN>=<valueN> where keys are alphabetically sorted.

Node

A node in a network that has a point site which is used in the calculation of the voronoi diagram of a cluster. Computes its own voronoi polygon from its list of neighbours. Does not store information on entire cluster.

OwnedKeyExpr

A Arc<str> newtype that is statically known to be a valid key expression.

QueryConsolidation

The replies consolidation strategy to apply on replies to a get.

Sample

A zenoh sample.

Selector

A selector is the combination of a Key Expression, which defines the set of keys that are relevant to an operation, and a set of parameters with a few intendend uses:

Session

A zenoh session.

Value

A zenoh Value.

ZenohId

The global unique id of a zenoh peer.

keyexpr

A str newtype that is statically known to be a valid key expression.

Enums

CongestionControl

The kind of congestion control.

ConsolidationMode

The kind of consolidation.

Encoding

The encoding of a zenoh zenoh::Value.

KnownEncoding

NodeStatus

Priority

The Priority of zenoh messages.

QueryTarget

The zenoh::queryable::Queryables that should be target of a zenoh::Session::get().

Reliability

SampleKind

The kind of a Sample.

Traits

IntoCallbackReceiverPair

A type that can be converted into a [Callback]-receiver pair.

Parameter

Parameters

A trait to help decode zenoh selector parameters.

Resolve

Zenoh’s trait for resolving builder patterns.

SessionDeclarations

Functions to create zenoh entities with 'static lifetime.

SplitBuffer

A trait for buffers that can be composed of multiple non contiguous slices.

SyncResolve

ValidatedMap

Type Definitions

ZInt

A zenoh integer.